rename YI to yi
